OnlineBankingWebApplicationProject
MODULE DESCRIPTION
The web site will be used by three different types of users: customers,
employees, and administrators.  The web site will provide a different
interface for each of the three types of users.
Customer interface
Account summary: This feature will allow a customer to view the
information related to all the accounts that he/she holds with the bank.
The information displayed will include the account ID, account type,
balance, and status of all the accounts that the customer holds with the
bank. The status of an account of an account can be either activated or
deactivated.
View statement:  This feature will allow a customer to view the details
of all transactions performed on any of their accounts held in the bank.
To obtain this information, customers would be required to specify their
account ID and the range of dates for which they want to view
transactions.
Transfer funds:  This feature will allow a customer to transfer money
from any of their accounts to another account in Westside bank.
Request for cheque book: This feature will allow a customer to make a
request for a cheque books. It will also allow a customer to view the
status of all previous requests for cheque books.
Apply for loan:  This feature will allow a customer to apply for a loan.
Before applying for a loan, the user will be able to calculate the EMI
for the loan. The feature also allows a customer to view the status of
all previous loan requests.
Change password: This feature will allow customers to change their
passwords.
Employee interface
Process loan requests: This feature will enable the employees to view
loan requests submitted by customers.  It will also enable an employee
to either approve or reject a loan
3
request.  If the loan is rejected, the employee will be prompted to
enter the reason for rejection.
Process cheque book requests: This feature will enable the employees to
view all requests for cheque book made by customers. It will also enable
an employee to specify whether a cheque book has been dispatched or not.
Change password: This feature will allow employees to change their
passwords.
Administrator interface
Create users: This feature will allow an administrator to create new
accounts for users and assign role to them.  For example, after creating
a login account for a customer, the administrator would be required to
assign the customer role to the user. The user name and password will be
sent to the customer through registered mail. In addition to creating
login accounts for customers, these features will also allow an
administrator to create login accounts for employees and others
administrators.
Manage users:  This feature will allow an administrator to mange users
by performing tasks such as assigning roles to users and modifying
user’s information.  The administrator will be able to modify
information such as address and phone number of the user. In addition,
the administrators will able to activate or deactivate a user login
account.
Modify user accounts: This feature ill allow the administrator to search
for a user name and assign a unique customer ID to the user name. This
customer ID will be obtained from the database that is maintained by the
existing computerized system for the bank.  In the existing database,
each customer ID is assigned one or more bank accounts. When a customer
ID is assigned to a user name, a person who logs in with this user name
will be able to access all bank accounts for the specified customer ID.
Modify accounts information: This feature will allow the administrator
to modify information, such as minimum balance and interest rate, for a
specific type of account. It will also enable the administrator to add
new type of accounts.

4
Modify loan information: This feature will allow the administrator to
modify information such as down payment and interest rate, for a
specific type of loan. It will also allow an administrator to add new
type of loan.
Update ticker: This feature will allow the administrator to create a new
news item that will be displayed in the news ticker on the home page.
The administrators will also be able to edit the news item and change
the status for a particular news item.  The status of a news item can be
either activated or deactivated. Only activated items will appear in the
news ticker. In addition, this feature will allow the administrator to
delete a particular ticker.
Change password: This feature will allow administrators to change their
passwords.